Refactor import paths in multiple files to use consistent casing for MyComponents directory and add new components including ApplicationTable, Button, CustomButton, Header, KPICards, Loader, and others for enhanced UI functionality.
