+++ src/MyComponents/Splash.tsx
