+++ src/MyComponents/SplashScreen.tsx
